improve some GUI elements in combination with macro tab
There are some contraints in using the TAB area for configuration.
because of too many elements neccessary for the macro GUI,
The font size had been set to 12px fix.

If you find that is to small, you can change it by installing an
own stylesheet file and calling ckb with
ckb -stylesheet=<stylesheet-file>

The complete TAB area (better the surrounding areas) should be
redesigned...