solves conflict between macros with long delays and color-updates
When using long macros with very long delays
(eg delays recorded while you typed very slowly)
and there is parallel a color-update necessary,
you can get an error in usb.c
(device closed as the keyboard resets itself).

Reason for this is misuse of the usb protocol
(sending two commands/request to the KB instead of
alternating request/reply)
because of missing synchronization between the two threads.

So we need a sync mechanism between these threads.
The mechanism is an additional mutex (mmutex).

The disadvantage of that synchronization is that
updating the color information is delayed
until all macros have been processed
(this also applies to type-ahead with several macros).
In the case of rather short macros or those with small delays,
this is not noticeable; in the case of wave motions,
the effect is occasionally recognizable.

And yes, we should rewrite ckb-daemon...