Key properly signed and reference to productive download file
Two major changes are included in the commit:

- The key was signed by me.
  It should also be signed by mattanger and possibly others.
  Caution: The key must not be exported as ASCII file, but only as a binary file!

- To test the function, the signed FIRMWARE file was originally uploaded to a branch of mine:
  https://raw.githubusercontent.com/frickler24/ckb-next/issues-26-Firmware-Incident/FIRMWARE
  (in line 36, kbfirmware.cpp).

  In order for the mechanism to go productive,
  the FIRMWARE file of the MASTER branch must be updated!

This patch must first be used in master branch so that the file and the key can be loaded at runtime.
Then the patch can be added to the testing branch and others.

If the file or the key in the master branch is not up-to-date,
an error message is displayed when the ckb client is started
(as an example, here a possible error message):

  Gpg: Signature from Thu Jun 30th 2016 09:15:49 CEST using RSA key ID 8DC8D309
  Gpg: [do not know]: invalid packet (ctb = 2d)
  Gpg: keydb_search failed: Invalid package
  Gpg: Can not check signature: Public key not found