improve the behavior when modifying modes for RGB keyboards
Changing the operating modes for the RGB keyboards that provide this feature
(such as the K70RGB or the K95RGB) will result in massive disruptions
to the daemon and the keyboard itself.

A test matrix is stored in the corresponding issue # 48.

The settings here refer to the timeout before sending a USB message.
The timeout was previously 10ms and has been doubled.

Minor corrections and the clipping of the debug output
in a #define DEBUG were also committed here.