Find more reasons for lack of writability of the config file
At the suggestion of @Tatokis, not only the writability of the
respective config file itself is evaluated,
but also reasons why the file might not be created.

This means that the standard function of QSessions
can no longer be used.

So attempts are first made to create a test entry in the Config.
If this test entry can be read after closing and
reopening the config file, the file is writable or can be created.