Find more reasons for lack of readability of the config file
At the suggestion of @tatokis,
not only the writability of the config file itself is evaluated,
but also reasons why the file could not be created.

This means that the original functions of QSessions
can no longer be used.
Attempts are first made to create a test entry in the Config.
If this test entry can be read after closing and reopening
the config file, the file is writable or can be created.