divided commenting code from changing log info
In this branch we just want to add comments to the code.
If there is something pure to correct the code
(as setting {}-pairs to clean up, declaring functions or
variables as static...), it will be done in this branch.

Additional debugging info or handling of return codes
is done in a separate branch.
Up to the last commit, these two things were intermixed.
With this commit, all code changes were eliminated.