enable mouse triggered keyboard macros
The first implementation of the macros was able to record macros, which
were to be executed on a key press.
Macros, which should be executed due to mouse clicks or special mouse
buttons, had to be created by cut & paste.
This commit also allows you to manage keyboard input for any mouse
button. The mechanism and the GUI are the same as in the keyboard
macros.

At the same time, some errors from the previous implementation were also
fixed. Also the macro storage was prepared for the future use of delay
data.