implement additional ways of informational messages
Standard function in Qt is to decide between 5 modes:

- qFatal() is used if you cannot do anything more than exiting. This
method needs a special handler and I think we should not use it, because
of handling notify-channels etc.
- qCritical() is for really critical situations (eg unknown usb events,
closed pipes, no more memory, SIGSEGV and so on). I Think we have some
error-handling in the code (some handling may be anhanced :-)) to use
this mode.
- qWarning() is less than critical and may be used if we could not load
new firmware etc.
- qInfo() is less again and might be used for messages like our standard
message "Downloaded new firmware list. 12 entries found."
- qDebug() should be used for debugging only. Here we can produce lots
of raw data for our purposes.