Correct typos and change qDebug into qCritical or qInfo
Some typos in the comments and the debug messages are fixed.

The user messages to stdout/stderr are now distinguished according to
their importance (qInfo, qWarning, qDebug, qCritical).
What is not yet used is qFatal because this requires a different form of
parameterization.

In rebindwidget.ui there was a part from a different patch,
this is reset to the original status.