Use a local file to bypass the cache mechanism in Qsettings
Because of internal (undocumented?) caching, the tests
with writing a test-setting, closing the QSettings and reopen it,
is not a secure test. A non-writable directory which owns another user
and does grant write access to this user is not detected.

Because of this an explicit separate file is written into the
settings-directory and removed afterwards
in addition to the older test variants.

This means that this test is not machine independant anymore.