Perform sync between macro and colorization with thread list
The first implementation for macros with delay management
has stopped the colorization while playing the macro content.

This Version handles the macro playing different:
It starts a single thread for each macro when
the appropriate key is pressed.
All tasks are self-syncing via a linked list (FIFO).

One Todo is left: When playing macros, all other keyboard input
except modifyer keys is ignored. A \todo is set as comment
where to heal this if anyone finds out...

With the new coding some comments for existing code
and changes for better readability are made and typos
have been fixed.